Unify lerp functions
The new lerp function is generic and takes a reference.
It's a bit unfortunate to have to AsRef() floats to call into lerp(),
but it's better than duplicate implementations. And that's a matter
of syntax, rather than an performance issues - compiler should know
the best way to get those primitives in there.