Create Normal trait, impl for Normal3i
I think we will do this for all of Vector3, Vector2, Point3,
Point2, and Normal3 types.

I was struggling a bit with how we would do combinations of
the Normal and Vector operations dot() and cross(), but I think
the associated type solution will make it work well.
We won't have many default implementations for these new traits
I think because of different algorithms for integers and floats,
but it's good to have the common trait available for e.g. bounds,
which would have something like Bounds2<T: Point2>.