Add boudns expand() functions
This and a few previous commits improve how the Tuple
classes handle generics, since bounding_box uses generic
points and vecs. I think we're reaching a pretty good spot with
vecmath.

The worst offense in VecMath is now I think that there's some
repetition in implemeting ops. For now that's fine, but
we could improve that with macros in the future. I wanted to do
some blanket implementations of ops with bounds on e.g. Point2
trait, but it violates some rules of the compiler. I think the
idiomatic way to handle that IS macros, so we can do that later.