Matrix-Vector multiplication
Generic function for multiplying square matrices and vectors.

It might be better if we had a const generic on V such that
our type system can check they're the same length.

This is a generic fucntion but we could use it to impl Mul for
various types if we want.