Use SpectrumI and NewTypes for variant enums
Use a combination SpectrumI trait and Spectrum enum to organize
the various spectrum types. This lets us implement functions for
only one function variant e.g. and use variants as solo types
if desired, while still being able to use the Enum for
runtime polymorphism when requested.

If this is too wordy for some, we could have used enum_dispatch
to generate the Spectrum enum automatically. But for clarity,
we'll go with this approach for now.