Set up get_named_spectrum()
We will use an Enum and a Match rather than a hashmap.
We will further lazily evaluate any spectra via once_cell.
This differs from PBRTv4, which has an Init() function to initialize
all the named spectra.
once_cell is making its way into std rust; I think this is pretty
idiomatic. It saves needing to compute a hash on a string for each
access compared to a hashmap, and the compiler is quite good
at optimizing the match.

We'll need to populate this with actual data, though.