Finish RgbIlluminantSpectrum
Note that the Illuminants are stored in an Arc; this seems a bit early
because we're not multithreaded yet - why not use Rc?

Well, it will become Arc eventually, and must be Arc for Spectrum
to satisfy send/sync for where we use Spectrum in once_cell sync'd
lazy evals, which we know will be necessary in the future.