Remove large clone() of Light
Uses Rc instead to avoid a large clone. Clone was large
due to Vec in DenselySampledSpectrum.
In the future we might improve this further by caching the
DenselySampledSpecturm in Light instead and storing only a ref,
in which case we could make Lights impl Copy and take the Lights
outside of the Rc again and just copy instead.