WIP Write basic BVH test, uncover bug
I think we'll want to use MaybeUninit to resolve this.
The alternative would be to restructure the code so that ordered_prims
is built as we go, rather than allocated all at once, but that's
probably less performant. We could also intialize it to some value
and just *change* it later, but containing an Rc<Primitive>,
it's a bit odd to have some "default primitive" as we'd need
to choose a variant.