Fix pre-allocating vectors
Uses Option to have "uninitialized" values in BVH construction be None,
and initializing to Some when they are initialized.
This has some runtime overhead (unless the compiler is smart enough
to optimize it away), but makes the code quite clear and lets us
do some runtime checks on the validity of the BVH construction.
An alternative would be to use MaybeUninit, which is like Option
but unsafe and without the checks.

For now, let's go with the Rusty Option approach, since we haven't
shown a performance need for MaybeUninit. Philosophical differences...

This only affects BVH construction, so it might not be a big deal
in any case.