Render minimal scene
This renders a scene with a small sphere acting as an area light
in an empty void.

I'm sure it's not perfect but - we've got a shape on the scene!
Development should be a lot more fun as we can see visual results.
Or more painful, as I uncover all the bugs.