Use Arc for colorspaces
This avoids expensive clone() operations.

We technically don't need Arc at the moment; we could use Rc.
But we're using sync::Lazy, which requires Arc instead of Rc.
We could use non-sync Lazy, but we *will* parallelize this, so
we can just leave it be so we don't need to switch back later