Adjust t_max in ray transforms
This was missing prior. It doesn't fix our occlusion problem, but
should be more correct.

Really, this only effects TransformedPrimitive right now -
triangles operate in render space (not transforming the ray),
and spheres are transforming the origin and direction separately.
Cameras involve ray transforms with times, but that's really for
animated transforms.