WIP Tokenizing, Parsing
Tokenizing and parsing; the things leading into the
BasicSceneBuilder.

Borrows code from the pbrt4 crate, which provides partial
parsing for .pbrt files. We don't use it directly because
it is not complete, and because of compatability reasons;
but it's a great starting point for the tokenizing and
parsing logic. Its license was added to the repo.

Also organizes some modules into the loading module.