Set target-cpu=native
This allows for rustc to generate a binary with optimizations for
the compiling/host machine. For us, this largely optimizes the
very critical fma instructions.

This is principally useful for development; in the long run,
we'd need builds for any target architecture being distributed to.
But we can stick with native for now since we're not distributing
this binary anytime soon.

In a very not-scientific experiment, got roughly a 2x speedup
on the killeroo-simple-dielectric scene. It's still quite slow,
however, compared to the PBRTv4 implementation on the same scene.
So there is a lot of room for improvement, I just need to examine
where it is. Will investigate an apaprent correctness issue, which
might reveal a related performance issue.