Wrap film in Mutex, dont accumulate samples
Rather than accumulating samples and procesing them sequentially,
which was taking up large amounts of memory and resulting in slower
renders, add samples to the film as we create them by wrapping the
film in a Mutex.

This DOES fix the issue of increasing memory consumption.
However, it slows down render speeds due to waiting on the lock()
for the film for each sample.

Possible solutions might be to accumulate samples for each sample
and only lock the film once for the entire tile.
Though, we also have the get_camera_sample() call within
evaluate_pixel_sample() which locks the film, which also
causes slowdowns.

We might instead forego the mutex and use get_mut_unchecked().
We would not satisfy the guarantee for safety, but because
the tiles should operate on different parts of the film plane,
we ~shouldn't~ be writing to the same memory.
I don't think that results in UB.